Speeding up test compilation by re-using the main executable compile steps
The previous version of the code recompiled all of the main executable files because they were linked as part of the source code. Now the core modules, everything except for `xml_converter.cpp` are compiled into a library first, then that library is used when linking against the main executable and the test framework executable.